+++ sqlobject/mysql/mysqlconnection.py
@@ -1,19 +1,28 @@
 from sqlobject import col, dberrors
-from sqlobject.compat import PY2
+from sqlobject.compat import PY2, string_type
 from sqlobject.converters import registerConverter, StringLikeConverter
 from sqlobject.dbconnection import DBAPI
 
 
 class ErrorMessage(str):
     def __new__(cls, e, append_msg=''):
-        if len(e.args) > 1:
-            obj = str.__new__(cls, e.args[1] + append_msg)
+        if e.__module__ == 'cymysql.err':
+            if isinstance(e.errmsg, string_type):
+                errmsg = e.errmsg
+            else:
+                errmsg = e.errmsg.reason
+            errcode = e.errno
         else:
-            obj = str.__new__(cls, append_msg)
-        try:
-            obj.code = int(e.args[0])
-        except ValueError:
-            obj.code = e.args[0]
+            if len(e.args) > 1:
+                errmsg = e.args[1]
+            else:
+                errmsg = ''
+            try:
+                errcode = int(e.args[0])
+            except ValueError:
+                errcode = e.args[0]
+        obj = str.__new__(cls, errmsg + append_msg)
+        obj.code = errcode
         obj.module = e.__module__
         obj.exception = e.__class__.__name__
         return obj
@@ -59,6 +68,16 @@ class MySQLConnection(DBAPI):
                         self.CR_SERVER_LOST = \
                             MySQLdb.constants.CR.CR_SERVER_LOST
                     self.ER_DUP_ENTRY = MySQLdb.constants.ER.DUP_ENTRY
+                elif driver == 'cymysql':
+                    import cymysql
+                    import cymysql.constants.CR
+                    import cymysql.constants.ER
+                    self.module = cymysql
+                    self.CR_SERVER_GONE_ERROR = \
+                        cymysql.constants.CR.CR_SERVER_GONE_ERROR
+                    self.CR_SERVER_LOST = \
+                        cymysql.constants.CR.CR_SERVER_LOST
+                    self.ER_DUP_ENTRY = cymysql.constants.ER.DUP_ENTRY
                 elif driver in ('connector', 'connector-python'):
                     import mysql.connector
                     self.module = mysql.connector
@@ -90,7 +109,7 @@ class MySQLConnection(DBAPI):
                     raise ValueError(
                         'Unknown MySQL driver "%s", '
                         'expected mysqldb, connector, connector-python, '
-                        'pymysql, mariadb, '
+                        'pymysql, cymysql, mariadb, '
                         'odbc, pyodbc or pypyodbc' % driver)
             except ImportError:
                 pass
@@ -230,7 +249,7 @@ class MySQLConnection(DBAPI):
                 conn.autocommit = auto
 
     def _force_reconnect(self, conn):
-        if self.driver == 'pymysql':
+        if self.driver in ('pymysql', 'cymysql'):
             conn.ping(True)
             self._setAutoCommit(conn, bool(self.autoCommit))
             if self.dbEncoding:
@@ -256,7 +275,7 @@ class MySQLConnection(DBAPI):
         # reconnect flag must be set when making the connection to indicate
         # that autoreconnecting is desired. In MySQLdb 1.2.2 or newer this is
         # done by calling ping(True) on the connection.
-        # PyMySQL needs explicit reconnect
+        # [PC]yMySQL need explicit reconnect
         # each time we detect connection timeout.
         for count in range(3):
             try:
@@ -268,7 +287,7 @@ class MySQLConnection(DBAPI):
                         raise dberrors.OperationalError(ErrorMessage(e))
                     if self.debug:
                         self.printDebug(conn, str(e), 'ERROR')
-                    if self.driver == 'pymysql':
+                    if self.driver in ('pymysql', 'cymysql'):
                         self._force_reconnect(conn)
                 else:
                     raise dberrors.OperationalError(ErrorMessage(e))